Move remaining LRU classes into the LRU module.
