+++ relstorage/cache/lru.py
@@ -21,31 +21,6 @@ Segmented LRU implementations.
 
 from .ring import Ring
 
-from cffi import FFI
-import os
-this_dir = os.path.dirname(os.path.abspath(__file__))
-
-# ffi = FFI()
-# ffi.cdef("""
-# typedef struct CPersistentRing_struct {
-#     ...;
-# } CPersistentRing;
-
-# typedef struct RSLRUEntry_struct {
-# 	CPersistentRing ring_entry;
-# 	uint_fast64_t frequency;
-# 	uint_fast64_t len;
-# } RSLRUEntry_t;
-
-# """
-# )
-
-# _FFI_RING = ffi.verify("""
-# #include "lru.h"
-# """, include_dirs=[this_dir])
-
-# ffi_new = ffi.new
-
 from .ring import ffi
 ffi_new = ffi.new
 ffi_new_handle = ffi.new_handle
@@ -55,6 +30,8 @@ from .ring import _FFI_RING
 
 _lru_update_mru = _FFI_RING.lru_update_mru
 _ring_move_to_head_from_foreign = _FFI_RING.ring_move_to_head_from_foreign
+_lru_probation_on_hit = _FFI_RING.lru_probation_on_hit
+
 
 class SizedLRURingEntry(object):
 
@@ -168,3 +145,108 @@ class SizedLRU(object):
             'size': self.size,
             'count': len(self._ring),
         }
+
+
+class EdenLRU(SizedLRU):
+
+    def __init__(self, limit, probation_lru, protected_lru, entry_dict):
+        SizedLRU.__init__(self, limit)
+        self.probation_lru = probation_lru
+        self.protected_lru = protected_lru
+        self.entry_dict = entry_dict
+
+    def add_MRU(self, key, value):
+        new_entry = SizedLRU.add_MRU(self, key, value)
+
+        if not self.over_size:
+            return new_entry
+
+        dct = self.entry_dict
+        probation_lru = self.probation_lru
+        protected_lru = self.protected_lru
+
+        if not probation_lru and not protected_lru.over_size:
+            # This is a modification of the algorithm. When we start out
+            # go ahead and populate the protected_lru directly
+            # from eden; only when its full do we start doing the probationary
+            # dance. This helps mitigate any issues with choosing segment sizes;
+            # we're going to occupy all the memory anyway, why not, it's reserved for us,
+            # so go ahead and fill it.
+
+            while self.over_size:
+                eden_oldest = self.get_LRU()
+                if eden_oldest.key is key:
+                    break
+
+                if eden_oldest.len + protected_lru.size > protected_lru.limit:
+                    # This would oversize protected. Move it to probation instead,
+                    # which is currently empty, so there's no need to choose a victim.
+                    # This may temporarily oversize us.
+                    probation_lru.take_ownership_of_entry_MRU(eden_oldest)
+                    break
+                else:
+                    protected_lru.take_ownership_of_entry_MRU(eden_oldest)
+
+            return new_entry
+
+        while self.over_size:
+            eden_oldest = self.get_LRU()
+            if eden_oldest.key is key:
+                break
+
+            #assert eden_oldest.__parent__ is None
+            #assert eden_oldest._Persistent__ring is None
+
+            if probation_lru.size + eden_oldest.len < probation_lru.limit:
+                # Cool, we can keep it.
+                probation_lru.take_ownership_of_entry_MRU(eden_oldest)
+                #assert eden_oldest.__parent__ is probation_lru
+            else:
+                # Snap, somebody has to go.
+                try:
+                    oldest_main_ring = probation_lru.get_LRU()
+                except (StopIteration, KeyError):
+                    # probation ring is empty, nothing to eject. This must be a large
+                    # item. Well, just accept it then to match what we used to do.
+                    probation_lru.take_ownership_of_entry_MRU(eden_oldest)
+                    continue
+
+                if oldest_main_ring.frequency > eden_oldest.frequency:
+                    # Discard this entry, it loses
+                    # print("Completely evicting item", oldest.key,
+                    #       "because main ring item", oldest_main_ring.key,
+                    #       "has better frequency", oldest_main_ring.frequency, oldest.frequency)
+                    self._ring.delete(eden_oldest)
+                    del dct[eden_oldest.key]
+                else:
+                    # eden item is more popular, keep it
+                    probation_lru.remove(oldest_main_ring)
+                    del dct[oldest_main_ring.key]
+
+                    probation_lru.take_ownership_of_entry_MRU(eden_oldest)
+                    #assert eden_oldest.__parent__ is probation_lru
+            self.over_size = self.size > self.limit
+        return new_entry
+
+class ProtectedLRU(SizedLRU):
+    pass
+
+
+class ProbationLRU(SizedLRU):
+
+    promote_count = 0
+    demote_count = 0
+    remove_count = 0
+
+    def __init__(self, limit, protected_lru, entry_dict):
+        SizedLRU.__init__(self, limit)
+        self.protected_lru = protected_lru
+        self.entry_dict = entry_dict
+
+    def on_hit(self, entry):
+        # Move the entry to the Protected LRU on its very first hit;
+        # It will become the MRU there.
+        protected_lru = self.protected_lru
+        protected_lru.over_size = _lru_probation_on_hit(self._ring.ring_home,
+                                                        protected_lru._ring.ring_home,
+                                                        entry.cffi_ring_node)
